Move viewmodel to separate file
